Create sampling_weights
Sampling weights were introduced to allow for differential weighing of individuals based on the probability of inclusion in the sample. 
All cases, as well as their siblings born within the cohort years, were included, resulting in the P=1. 
There were 270,799 children born in the cohort years in the Meuhedet records. Among those, XXX (XX%) were sampled at random, i.e. the baseline probability of being included in the sample was 0.XX% for singletons. Given that information was also collected on all siblings of the members of the subcohort that were born between 1997 and 2008, this probability increased for families with more than one child originally selected for the cohort. For example, for families with two children in the main cohort, children would be included in the study if either of them was selected. Given their inclusion was not mutually inclusive, the probability with which they were included amounted to the sum of probabilities for each child, minus the probability of their joint inclusion. Consequently, the more children from a given family were included in the sample, the higher the probability that either of them would contribute data to the study. 
All siblings of the children in the main cohort, who were also born within the cohort years, were by default included in the full subcohort data. Their inclusion / number, did not, howerver, affect the probability of inclusion of data from a given family, and the probabilty of their sampling was determined solely on the number of their siblings included in the original selection. Given that they were included always if their siblings were, their sampling probability was the same as that of their siblings in the main cohort.